refactor(reporter): split report_generator.py into focused modules
report_generator.py was 2748 lines mixing HTML template rendering, chart
data prep, JSON/CSV serialization, and shared metrics helpers. Split into:

- report_generator.py   (220 lines): generate_report() dispatcher + JSON
                                      + CSV + test-compat re-exports
- report_data.py        (136 lines): metrics helpers shared across formats
                                      (coverage_pct_of_min_hourly, etc.)
- chart_data.py         (261 lines): timeseries -> chart data + optimal
                                      coverage computation
- html_sections.py      (527 lines): self-contained HTML fragments
                                      (breakdown table, active plans table,
                                      spike guard banner, tabs, scheduler
                                      preview rows)
- html_report.py       (1590 lines): the main HTML template

Public surface unchanged: generate_report() + the private names
test_report_generator_units.py imports (_build_breakdown_table_html,
_get_type_metrics_for_report, _parse_plan_dates,
_render_spike_guard_warning_banner) are re-exported from
report_generator.py so the test file doesn't need changes.

No behavior change — extracted functions are byte-for-byte copies with
private underscore stripped from cross-module helpers.
